requirements.txt
av==12.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.0.3
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.1.3
onnxruntime==1.20.0
packaging==24.2
protobuf==5.28.3
pycparser==2.22
PyYAML==6.0.2
requests==2.32.3
setuptools==75.3.0
sounddevice==0.5.1
sympy==1.13.3
tokenizers==0.20.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
